bug de Joshua and Amy corrigé
"I am using ledmac+ledpar to typeset parallel text as part of my research on the Jarai language. I'd like to use endnotes rather than footnotes. However, the notes from the left column won't print if I call them after the right column is complete (and they don't even show up in the .end file). If I call the A series (used on the left) between Leftside and Rightside, then the endnotes do show up, but not in the right place."